add Pixel set up
