+++ DuckDuckGo.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -33,7 +33,7 @@
       "location" : "https://github.com/DuckDuckGo/BrowserServicesKit",
       "state" : {
         "branch" : "sabrina/experiment-pixels",
-        "revision" : "c62fac4c79969c47c5396b23686c400df30b48eb"
+        "revision" : "8db91da8441a247dece2bfe8ba9f0db0fbbf1eeb"
       }
     },
     {
